[ALOY-937] Account for platform-specific folders when copying widget assets

+++ Alloy/commands/compile/compilerUtils.js
@@ -518,21 +518,41 @@ exports.inspectRequireNode = function(node) {
 	};
 };
 
-exports.copyWidgetResources = function(resources, resourceDir, widgetId) {
+exports.copyWidgetResources = function(resources, resourceDir, widgetId, opts) {
+	opts = opts || {};
 	_.each(resources, function(dir) {
 		if (!path.existsSync(dir)) { return; }
+		logger.trace('WIDGET_SRC=' + path.relative(compilerConfig.dir.project, dir));
 		var files = wrench.readdirSyncRecursive(dir);
 		_.each(files, function(file) {
 			var source = path.join(dir, file);
+
+			// make sure the file exists and that it is not filtered
+			if (!fs.existsSync(source) ||
+				(opts.filter && opts.filter.test(file)) ||
+				(opts.exceptions && _.contains(opts.exceptions, file))) {
+				return;
+			}
+
 			if (fs.statSync(source).isFile()) {
-				var destDir = path.join(resourceDir, path.dirname(file), widgetId);
+				var dirname = path.dirname(file);
+				var parts = dirname.split(/[\/\\]/);
+				if (opts.titaniumFolder && parts[0] === opts.titaniumFolder) {
+					dirname = parts.slice(1).join('/');
+				}
+
+				var destDir = path.join(resourceDir, dirname, widgetId);
 				var dest = path.join(destDir, path.basename(file));
 				if (!path.existsSync(destDir)) {
 					wrench.mkdirSyncRecursive(destDir, 0755);
 				}
+
+				logger.trace('Copying ' + file.yellow + ' --> ' +
+					path.relative(compilerConfig.dir.project, dest).yellow + '...');
 				U.copyFileSync(source, dest);
 			}
 		});
+		logger.trace(' ');
 	});
 };
